# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AICommandRequest:
    properties:
      command:
        type: string
      input:
        type: string
    required:
    - command
    - input
    type: object
  models.AICommandResponse:
    properties:
      command:
        type: string
      result:
        type: string
      session_id:
        type: string
    type: object
  models.AISession:
    properties:
      command:
        type: string
      created_at:
        type: string
      id:
        type: string
      input_summary:
        type: string
      output_summary:
        type: string
      user_id:
        type: string
    type: object
  models.AuthResponse:
    properties:
      expires_in:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.ColumnSummary:
    properties:
      count:
        type: integer
      name:
        type: string
      task_ids:
        items:
          type: string
        type: array
    type: object
...
  models.DashboardStats:
    properties:
      columns:
        items:
          $ref: '#/definitions/models.ColumnSummary'
        type: array
      tasks_by_priority:
        additionalProperties:
          type: integer
        type: object
      tasks_by_status:
        additionalProperties:
          type: integer
        type: object
    type: object
  models.FocusSummary:
    properties:
      completed_minutes:
        type: integer
      remaining_minutes:
        type: integer
      total_minutes:
        type: integer
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  models.SignupRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.Task:
    properties:
      column_name:
        type: string
      created_at:
        type: string
      description:
        type: string
      estimated_minutes:
        type: integer
      id:
        type: string
      position:
        type: integer
      priority:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.TimeBlock:
    properties:
      completed:
        type: boolean
      created_at:
        type: string
      date:
        type: string
      end_time:
        type: string
      id:
        type: string
      start_time:
        type: string
      task_id:
        type: string
      user_id:
        type: string
    type: object
  models.TodayView:
    properties:
      date:
        type: string
      focus_summary:
        $ref: '#/definitions/models.FocusSummary'
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      time_blocks:
        items:
          $ref: '#/definitions/models.TimeBlock'
        type: array
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_confirmed:
        type: boolean
      id:
        type: string
      name:
        type: string
      supabase_id:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for task management with Kanban + Timeboxing + AI commands
  title: FLOWSTATE API
  version: "1.0"
paths:
  /ai/command:
    post:
      consumes:
      - application/json
      description: |-
        Executes AI commands for workflow automation:
        - /descomponer: Break a task into subtasks
        - /estimar: Estimate time for a task
        - /planificar: Create a daily plan
        Maintains conversation context (last 5 sessions) for personalized responses.
        Auto-compacts context when 6+ sessions accumulated.
      parameters:
      - description: AI command and input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AICommandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AICommandResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: AI productivity assistant
      tags:
      - ai
  /ai/sessions:
    get:
      description: Obtiene el historial de sesiones de IA del usuario
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AISession'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Listar sesiones de IA
      tags:
      - ai
  /ai/sessions/{id}:
    get:
      description: Obtiene una sesión específica por ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AISession'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obtener sesión de IA
      tags:
      - ai
  /auth/confirm:
    get:
      description: Confirms user email with the provided token
      parameters:
      - description: Confirmation token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Confirm email
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns Supabase JWT
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /auth/me:
    get:
      description: Returns the profile of the currently authenticated user based on
        JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a new user account and sends confirmation email
      parameters:
      - description: Registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register new user
      tags:
      - auth
  /dashboard:
    get:
      description: 'Returns aggregated statistics: tasks grouped by status and priority,
        column summaries with task counts. Use to power dashboard visualizations and
        productivity analytics.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DashboardStats'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Kanban board overview
      tags:
      - dashboard
  /focus:
    get:
      description: 'Returns focus time analytics: total planned minutes, completed
        minutes, and remaining minutes. Use to display daily focus progress and productivity
        metrics.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FocusSummary'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Focus time metrics for today
      tags:
      - planning
...
  /today:
    get:
      description: Returns today's tasks and time blocks for planning. Provides focus
        time metrics comparing planned vs completed minutes. Use this endpoint to
        power the daily planning view.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TodayView'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Daily productivity snapshot
      tags:
      - planning
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"