# pkg/models/ai.go
package models

import "github.com/google/uuid"

type UpdateNoteRequest struct {
	ContentJSON string `json:"content_json" binding:"required"`
}

type AICommandRequest struct {
	Command string `json:"command" binding:"required"`
	Input   string `json:"input" binding:"required"`
}

type AICommandResponse struct {
	Command   string    `json:"command"`
	Result    string    `json:"result"`
	SessionID uuid.UUID `json:"session_id"`
}